Upgrade react-redux to 7.2.0
