+++ home/package.json
@@ -59,9 +59,9 @@
     "react-navigation-material-bottom-tabs": "1.1.1",
     "react-navigation-stack": "^2.0.15",
     "react-navigation-tabs": "^2.4.1",
-    "react-redux": "^6.0.1",
+    "react-redux": "^7.2.0",
     "reanimated-bottom-sheet": "^1.0.0-alpha.18",
-    "redux": "^4.0.1",
+    "redux": "^4.0.5",
     "redux-thunk": "^2.3.0",
     "semver": "^6.1.1",
     "sha1": "^1.1.1",
